refactor: 코드 정렬 및 패키지 변경
