+++ src/main/kotlin/bridge/Application.kt
@@ -1,5 +1,8 @@
 package bridge
 
+import bridge.view.InputView
+import bridge.view.OutputView
+
 fun main() {
     OutputView().printStart()
     println()
