+++ src/main/kotlin/bridge/view/MovingChecker.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.view
 
 import constant.Message
 import constant.Symbol
@@ -7,8 +7,8 @@ class MovingChecker(private val input: String?) {
     fun checkAll(): Boolean {
         try {
             checkValue()
-            } catch (e: IllegalArgumentException) {
-                println(Message.BASIC_ERROR.message + Message.MOVING_ERROR.message)
+        } catch (e: IllegalArgumentException) {
+            println(Message.BASIC_ERROR.message + Message.MOVING_ERROR.message)
             return true
         }
         return false
