+++ src/main/kotlin/bridge/view/OutputView.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.view
 
 import constant.Message
 import constant.Symbol
